Move options to util
